refactor : 모달을 재사용할 수 있게 구조 변경
